Optimizations here and here

+++ src/include/string.h
@@ -6,11 +6,11 @@
 #include <stddef.h>
 #include <stdint.h>
 
-void *memset(void *b, int c, int len);
+void *memset(void *b, int c, size_t len);
 unsigned int strlen(const char *s);
 char *strcat(char *s1, const char *s2);
 char *strcpy(char *destination, const char *source);
-char *itoa(int value, char *str, int base);
+char *itoa(int value, char *str, size_t base);
 int atoi(char *str);
 int strcmp(const char *input, const char *check);
 char tolower(char ch);
